Fixed SDRAM config
The SDRAM is configured at bootup (or reset) through the Device
Configuration Data (dcd). This would also be possible after code
execution but through this there's direct access to external SDRAM or
similar.

The problem was that the underlying SEMC controller was configured
incorrectly for the external SDRAM TQ uses as it assumed a 32bit "port
size" even though the W988D6FBGX6 has a 16bit I/O width.

A change to the last two bits of the SDRAMCR0 register helped out.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>