The led on the dev board does blinky stuff :D
there weren't a lot of changes between this and the other commit. The
only problem left was that the `SKIP_DCDC_ADJUSTMENT` flag is not
important. It is mentioned in the `clock_config.c` the underlying
function is never called. Instead the CPU bring up is done by
`zephyr/soc/arm/nxp_imx/rt/soc_rt11xx.c`. Here a `ADJUST_DCDC` is
mentioned.
Sadly this flag seems to be pinned to true in the soc Kconfig for the
iMXRT1176 soc. I'm not sure how to change an soc flag in the board config
so I've just adapted zephyrs soc Kconfig and it works now ^^

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>